#52 Implement HttpSession backed up by Redis datastore

+++ src/main/java/com/graffitab/server/api/authentication/UserSessionListener.java
@@ -1,18 +1,20 @@
 package com.graffitab.server.api.authentication;
 
+import com.graffitab.server.service.user.UserSessionService;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSessionEvent;
 import javax.servlet.http.HttpSessionListener;
 
-import lombok.extern.log4j.Log4j2;
-
-import org.springframework.stereotype.Component;
-
-import com.graffitab.server.service.user.UserSessionService;
-
 @Component
 @Log4j2
 public class UserSessionListener implements HttpSessionListener {
+
+	@Value("${session.backups.enabled:true}")
+	private Boolean sessionBackupsEnabled;
 
 	@Resource
 	private UserSessionService userSessionService;
@@ -26,11 +28,14 @@ public class UserSessionListener implements HttpSessionListener {
 
 	@Override
 	public void sessionDestroyed(HttpSessionEvent se) {
+
 		if (log.isDebugEnabled()) {
 			log.debug("Session destroyed: {}", se.getSession().getId());
 		}
 
-		userSessionService.deleteSession(se.getSession().getId());
+		if (sessionBackupsEnabled) {
+			userSessionService.deleteSession(se.getSession().getId());
+		}
 	}
 
 }
